% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_w.R
\name{gen_w}
\alias{gen_w}
\title{Simulate a symmetric binary spatial weights matrix}
\usage{
gen_w(n, cutoff = 0.2)
}
\arguments{
\item{n}{The number of observations to include}

\item{cutoff}{A threshold above which observations are no longer considered
neighbors. Observations are sampled in a (-1, 1) space. The default cutoff
is 0.2}
}
\value{

}
\description{
Simulate a symmetric binary spatial weights matrix
}
\examples{
gen_w(20)
}
