% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialhelpers.R
\docType{package}
\name{spatialhelpers}
\alias{spatialhelpers}
\title{spatialhelpers: A package containing convenience functions for working with
spatial data}
\description{
The spatialhelpers package currently provides functions for constructing and
modifying spatial weights matrices.
}
\author{
\strong{Maintainer}: Logan Stundal \email{stund005@umn.edu}(\href{https://www.loganstundal.com/}{loganstundal.com})
}
