% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_w.R
\name{tidy_w}
\alias{tidy_w}
\title{Tidy a pre-specified spatial weights matrix}
\usage{
tidy_w(formula, data, id, w)
}
\arguments{
\item{formula}{A model formula to identify appropriate observations for subsetting}

\item{data}{A data frame containing variables to be used in a spatial model}

\item{w}{A spatial weights matrix containing observations that
correspond to data rows with missing values}

\item{panel_id}{A unique ID variable that correpsonds to row and column names in W}
}
\value{
A subset w which has dropped rows and columns that have missing values on variables
in the supplied formula.
}
\description{
This function tidies a pre-specified spatial weights matrix. Using a unique ID
from a data frame as well as a model formula, this function subsets an existing W-matrix
by dropping rows and columns which correspond to missing values in the data.
}
\examples{
n   <- 20

dat <- data.frame(obs_id = 1:n, x1 = rnorm(n))
dat$x1[sample(1:20, size = 3)] <- NA

w   <- gen_w(n)
rownames(w) <- colnames(w) <- dat$obs_id

w2 <- tidy_w(formula = ~x1,
             data    = dat,
             id      = "obs_id",
             w       = w)
}
