% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us.R
\docType{data}
\name{us}
\alias{us}
\title{A simple feature collection of 48 contiguous states}
\format{
A data frame with 48 rows and 1 variable:
\describe{
\item{name}{State name}
\item{geometry}{Spatial geometry}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
us
}
\description{
A dataset containing geometry for 48 contiguous US states.
}
\examples{
data(trade)
head(trade)
}
\keyword{datasets}
