% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_borders.R
\name{w_borders}
\alias{w_borders}
\title{Compute distance based spatial weights}
\usage{
w_borders(data, id)
}
\arguments{
\item{data}{Spatial data of class sf}

\item{id}{A variable in the data uniquely identifying observations}
}
\value{

}
\description{
Compute distance based spatial weights
}
\examples{
data(us)
w <- w_borders(data = us, id = name)
}
