% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_data.R
\name{w_data}
\alias{w_data}
\title{w_data}
\usage{
w_data(
  data,
  id1,
  id2,
  flow1,
  flow2,
  flow,
  missing_list = FALSE,
  quitetly = FALSE
)
}
\arguments{
\item{data}{A data frame containing dyadic flows}

\item{id1}{Variable identifying unit 1}

\item{id2}{Variable identifying unit 2}

\item{flow1}{Flows into unit 1 from unit 2}

\item{flow2}{Flows into unit 2 from unit 1}

\item{flow}{Total flows between dyads. This will override \code{flow1} and
\code{flow2} if supplied and return a symmetric matrix}

\item{missing_list}{For instances where the data contains dyads with no
reported flows, the returned weights matrix will contain NAs. Setting this
argument to TRUE returns a list of these observations}

\item{quitetly}{Suppress warning messages}
}
\value{
A matrix consisting of weights derived from data
}
\description{
\code{w_data} constructs spatial weights using dyadic
data recorded in a data frame. These weights may be symmetric or
asymmetric depending on user preference.
}
\details{
Construct a spatial weights matrix using observations in a data frame
}
\examples{
data(trade)
head(trade)
# Construct a symmetric dyadic matrix
w_symmetric <- w_data(data = trade, id1 = importer1, id2 = importer2,
                      flow = total_trade)

# Recode missing values to 0 if these correspond to cases with no flows:
w_symmetric[is.na(w_symmetric)] <- 0

# Verify symmetry:
isSymmetric(w_symmetric)
w_symmetric["Canada", "Mexico"]
w_symmetric["Mexico", "Canada"]

# Return list of dyads with no flows to explore whether these are missing data
w_missing <- w_data(data = trade, id1 = importer1, id2 = importer2,
                    flow = total_trade, missing_list = TRUE)

w_missing$France # Dyads not reporting trade with France


# Construct an asymmetric dyadic matrix
head(trade)
w_asymmetric <- w_data(data  = trade, id1 = importer1, id2 = importer2,
                       flow1 = flow1, flow2 = flow2)

# Verify asymmetric
isSymmetric(w_asymmetric)

w_asymmetric["Canada", "Mexico"] # Imports of Canada from Mexico in current US millions $
w_asymmetric["Mexico", "Canada"] # Imports of Mexico from Canada in current US millions $
}
