% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_report.R
\name{w_report}
\alias{w_report}
\title{Report neighbors for an index in a spatial weights matrix}
\usage{
w_report(id, w, value = FALSE)
}
\arguments{
\item{id}{Index identifier for unit to report neighbors}

\item{w}{A spatial weights index}

\item{value}{Whether to report weight value, default is False}
}
\value{
A vector of neighbor names or recorded neighbor values
}
\description{
Report neighbors for an index in a spatial weights matrix
}
\examples{
data(us)
w <- w_borders(data = us, id = name)
w_report(id = "Minnesota", w = w, value = TRUE)
}
