% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_sim.R
\name{w_sim}
\alias{w_sim}
\title{Simulate a symmetric binary spatial weights matrix}
\usage{
w_sim(n, cutoff = 0.2)
}
\arguments{
\item{n}{The number of observations to include}

\item{cutoff}{A threshold above which observations are no longer considered
neighbors. Observations are sampled in a (-1, 1) grid space. The default
cutoff is a distance of 0.2}
}
\value{

}
\description{
Simulate a symmetric binary spatial weights matrix
}
\examples{
w_sim(20)
}
