% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_borders.R
\name{w_borders}
\alias{w_borders}
\title{Compute spatial weights based on shared border lengths}
\usage{
w_borders(data, id)
}
\arguments{
\item{data}{Spatial data of class sf}

\item{id}{A variable in the data uniquely identifying observations}
}
\value{
A matrix consisting of shared border lengths between units
}
\description{
Compute spatial weights based on shared border lengths
}
\examples{
data(us)
w <- w_borders(data = us, id = name)
}
